+++ test/integration/master/synthetic_master_test.go
